refactor(tests): extract shared fixtures to conftest.py
